requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asynctest==0.13.0
attrdict==2.0.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==19.10b0
boto3==1.33.13
botocore==1.33.13
braceexpand==0.1.7
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.0.2
colorama==0.4.6
comm==0.1.4
cycler==0.11.0
Cython==3.0.8
decorator==5.1.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
editdistance==0.6.2
exceptiongroup==1.2.0
fasttext==0.9.2
filelock==3.12.2
flake8==5.0.4
Flake8-pyproject==1.2.3
fonttools==4.38.0
frozendict==2.4.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.3
g2p-en==2.1.0
gdown==4.7.3
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.26.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
h5py==3.8.0
huggingface-hub==0.16.4
Hydra==2.5
hydra-core==1.3.2
idna==3.6
imagesize==1.4.1
importlib-metadata==4.2.0
importlib-resources==5.12.0
inflect==6.0.5
iniconfig==2.0.0
ipadic==1.0.0
ipython==7.34.0
ipywidgets==8.1.1
isort==4.3.21
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jupyterlab-widgets==3.0.9
kaldi-io==0.9.8
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.5
latexcodec==2.0.1
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.39.1
lxml==5.1.0
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mecab-python3==1.0.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nemo-toolkit==1.7.0
nltk==3.8.1
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.1
OpenCC==1.1.6
packaging==23.2
pandas==1.3.5
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.3
pathspec==0.11.2
pesq==0.0.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.0.0
pluggy==1.2.0
pooch==1.8.0
portalocker==2.7.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pybtex==0.24.0
pybtex-docutils==1.0.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.13
pyDeprecate==0.3.1
pydub==0.25.1
pyflakes==2.5.0
Pygments==2.17.2
pyparsing==3.1.1
pypinyin==0.50.0
PySocks==1.7.1
pystoi==0.4.1
pytest==7.4.4
pytest-runner==6.0.1
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2023.3.post1
pyupgrade==3.3.2
PyYAML==5.4.1
rapidfuzz==3.4.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.8.2
sacrebleu==2.4.0
sacremoses==0.0.53
safetensors==0.4.1
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.99
sentry-sdk==1.39.2
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
sox==1.4.1
soxr==0.3.7
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.6.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tokenize-rt==5.0.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.12.1+cu116
torch-stft==0.1.4
torchaudio==0.12.1+cu116
torchmetrics==0.11.4
torchvision==0.13.1+cu116
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typed-ast==1.5.5
typer==0.9.0
typing_extensions==4.7.1
Unidecode==1.3.8
urllib3==1.26.18
wandb==0.16.2
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==2.2.3
wget==3.2
widgetsnbextension==4.0.9
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
youtokentome==1.0.6
zipp==3.15.0
